% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_reqs.R
\name{sys_reqs}
\alias{sys_reqs}
\title{Print SystemRequirements}
\usage{
sys_reqs(pkg)
}
\arguments{
\item{pkg}{CRAN Package name string}
}
\description{
This function prints the SystemRequirements string of the input package's
dependencies.
}
